Elevate the warning level to /W4 and fix warnings. Specific warnings disabled for `engextcpp.hpp` and `engextcpp.cpp` since they are provided by Microsoft. Added a precompiled header to disable warnings in `engextcpp.hpp`. The build is also much faster.

+++ extension.h
@@ -27,5 +27,5 @@ public: // Known structs.
 	void KnownStructObjectHandler(_In_ PCSTR TypeName, _In_ ULONG Flags, _In_ ULONG64 Offset);
 
 private:
-	std::vector<DEBUG_STACK_FRAME> getStackFrames(int numFrames = 1024);
+	std::vector<DEBUG_STACK_FRAME> getStackFrames(size_t numFrames = 1024);
 };
